jaylog-fastapi/requirements.txt
anyio==3.6.2
autopep8==2.0.0
bcrypt==4.0.1
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
mariadb==1.1.5.post3
MarkupSafe==2.1.1
mypy==0.991
mypy-extensions==0.4.3
orjson==3.8.2
packaging==21.3
pycodestyle==2.10.0
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.5.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
